Generics
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@13891 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/test/com/liferay/portal/service/persistence/BasePersistenceTestCase.java
@@ -51,7 +51,7 @@ public class BasePersistenceTestCase extends TestCase {
 	}
 
 	protected double nextDouble() throws Exception {
-		return (double)CounterLocalServiceUtil.increment();
+		return CounterLocalServiceUtil.increment();
 	}
 
 	protected int nextInt() throws Exception {
